+++ src/FilamentMcpServiceProvider.php
@@ -18,6 +18,7 @@ class FilamentMcpServiceProvider extends PackageServiceProvider
     {
         $package->name(static::$name)
             ->hasConfigFile()
+            ->hasViews()
             ->hasMigrations([
                 'create_filament_mcp_tokens_table',
                 'create_filament_mcp_tool_calls_table',